requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==22.0.0
psycopg2==2.9.9
SQLAlchemy==2.0.31